requirements.txt
wheel
Django==3.1
djangorestframework==3.11.1
django-cors-headers==3.4.0
python-decouple==3.3
pytest==6.0.1
pytest-django==3.9.0